+++ src/storm/models/sparse/Model.h
@@ -419,10 +419,28 @@ namespace storm {
                 boost::optional<std::shared_ptr<storm::storage::sparse::ChoiceOrigins>> choiceOrigins;
                 
             };
-            
+
 #ifdef STORM_HAVE_CARL
+            /*!
+             * Get all probability parameters occurring on transitions.
+             * @param model Model.
+             * @return Set of parameters.
+             */
             std::set<storm::RationalFunctionVariable> getProbabilityParameters(Model<storm::RationalFunction> const& model);
+
+            /*!
+             * Get all parameters occurring in rewards.
+             * @param model Model.
+             * @return Set of parameters.
+             */
             std::set<storm::RationalFunctionVariable> getRewardParameters(Model<storm::RationalFunction> const& model);
+
+            /*!
+             * Get all parameters (probability and rewards) occurring in the model.
+             * @param model Model.
+             * @return Set of parameters.
+             */
+            std::set<storm::RationalFunctionVariable> getAllParameters(Model<storm::RationalFunction> const& model);
 #endif
         } // namespace sparse
     } // namespace models